+++ Sources/RNLibreToolsiOS13/LibreToolsResources/Libre2US.swift
@@ -1,0 +1,15 @@
+//
+//  Libre2US.swift
+//  LibreWrapper
+//
+//  Created by Yury Dymov on 4/2/22.
+//
+
+import Foundation
+
+final class Libre2US: AbstractLibre2 {
+    
+    override var type: SensorType {
+        return .libre2US
+    }
+}
