+++ Sources/RNLibreToolsiOS13/LibreToolsResources/LibreToolbox.swift
@@ -1,0 +1,13 @@
+//
+//  LibreToolbox.swift
+//  LibreWrapper
+//
+//  Created by Yury Dymov on 4/2/22.
+//
+
+import Foundation
+
+struct LibreToolbox {
+    let logger: Logging
+    let debugLevel: Int
+}
